requirements.txt
fastapi>=0.109.1
GPUtil>=1.4.0
psutil>=5.9.8
py-cpuinfo>=9.0.0
uvicorn